Improvement: avoiding gaps in user gid sequence.
The gid sequence was increased in every update of user/author
because we use the atomic operation "INSERT ... ON CONFLICT ...".
It is the known and normal behavior.
But, in that case, it is not required to be atomic.
So we avoid gaps checking previously if user exists
to minimize the conflict cases. Now the conflict only can appear on
concurrent user creations and then nothing is done for the second
(probably repeated request).