Now dataset accesses are not removed at the end,
they remain stored in db for the history
and just marked with the column "closed" changed to true.
So, the DELETE /datasetAccess/<id> operation is changed to PATCH
with some info required in the body.
Some additional info is also stored in db for the history
as new columns in the table.
Some of them are dedicated to the openchallenge
and the info will be provided by the k8s operator only if the special
response "openchallenge' is returned in POST /datasetAccessCheck.