Fix: metadata collected now from all the series.
Previously all metadata fields were obtained from the first dicom file of the first series,
and stored in the study table in DB.
But actually some of the fields are different for each series:
body part, modality and manufacturer.
So now the first dicom file of all series is readed
and that fields of metadata are stored in a new table "series" in DB.
The functions to obtain aggregated info of datasets (dataset.py)
are adjusted to include all the series.
And the EUCAIM search functions (storage.py) are adjusted also
to search in the new series tables.