Add series hash cache.
It is very useful during a global integrity check when there are
different datasets containing the same studies and series.
The hash of each series will be calculated only one time.
And during the creation of a dataset when some of the studies/series
selected are the same as in another datasets created before (recently).
Even during a relaunch of a dataset creation process if it has been
interrupted (the process will be resumed because the hash of first
studies/series will already be calculated and cached).