Improve response time for dataset creation.
Datasets are growing and currently big datasets creation in database took long time (several minutes),
so the response time for the POST /datasets was too much and gateway and client threw timeout error.

But now the more time consuming operations has been moved to the creation worker:
 - checks over datalake filesystem
 - creation of studies in DB
So the response time is normal again (seconds).