Enhance the users management.
New operation GET /users based on the Keycloak Admin API.
Extended the operation  GET /users/{username} with two scopes:
 - "gid" (the previous and default) to get only the gid,
 - "all" (the new one) to get all details,
    it includes all the relevant data from Keycloak:
    attributes, projects and roles (this last two extracted from groups).
The operation PUT /users/{username} has been accordingly adjusted
to set or update the projects and roles for a user in Keycloak.
Add operations to list all possible values for some of the properties
of the PUT (of user):
 - GET /userRoles, for the property "roles"
 - GET /userSites, for the property "siteCode"
 - And a new purpose "userManagement" in GET /projects,
   for the property "projects".

One important change is now users can be added without access to datasets
(e.g. data ingestion users) and GID assignment is not needed for that users,
so now a GID is not assigned to all users on creation, it can be null,
and only will be autogenerated when required with the "gid=-1".